requirements.txt
tensorflow-gpu==2.1.0
numpy
pandas
imgaug
pillow